+++ contracts/storage/AccountBalanceStorage.sol
@@ -20,5 +20,5 @@ abstract contract AccountBalanceStorageV1 {
     // first key: trader, second key: baseToken
     mapping(address => mapping(address => LibAccountMarket.Info)) internal _accountMarketMap;
     // Index price oracle underlying index
-    uint64 internal _underlyingPriceIndex;
+    uint256 internal _underlyingPriceIndex;
 }
